+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `on-failure` config to ack, reject or throw exception
 - Added `Pigeon::headers([])` and config `headers` key
 - Added possibility to use callable on headers config
+### Fixed
+- Fixed failing when not set fallback and throw exception
+- Fixed wrong `MessageProcessorTest` tests
 
 ## [1.2.0]
 ### Added